+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Improved stack-images CLI.
 * New crop-border CLI.
 * Add `contains` to `kwimage.Box`
+* Add `kwimage.adjust` to adjust simple image properties.
 
 ### Fix
 * `Detections.draw_on` no longer breaks if some classes are unknown (i.e. null)